Exercise 012
João Papo-de-Pescador, a good man, bought a microcomputer to monitor the daily income of his work. Every time he brings in a weight of fish greater than that established by the fishing regulations of the state of São Paulo (50 kilos), he must pay a fine of R$ 4.00 per excess kilo. João needs you to write a program that reads the variable weight (weight of fish) and calculates the excess. Record in the excess variable the amount of kilos beyond the limit and in the fine variable the amount of the fine that João must pay. Print the program data with the appropriate messages.