Exercise 013
Make a program that asks how much you earn per hour and the number of hours worked in the month. Calculate and show the total of your salary in that month, knowing that 11% are deducted for the Income Tax, 8% for the INSS and 5% for the union, make a program that gives us:

1. Gross salary.
2. how much you paid to the INSS.
3. how much you paid to the union
4. the net salary.
5. Calculate the discounts and the net salary, according to the table below:
6. + Gross Salary: BRL
7. - Income Tax (11%): BRL
8. - INSS (8%): ​​BRL
9. - Union (5%): BRL

= Net Salary: BRL

Obs.: Gross Salary - Discounts = Net Salary.